third_party/blink/web_tests/virtual/webrtc-wpt-plan-b/external/wpt/webrtc/no-media-call-expected.txt

This is a testharness.js-based test.
FAIL Can set up a basic WebRTC call with no data. assert_equals: expected (number) 1 but got (undefined) undefined
Harness: the test ran to completion.
